+++ node/target_image_utilities.h
@@ -1,10 +1,12 @@
 const uint16_t IMAGE_COUNT = 8186;
 
-void load_random_image(float *image_data_buffer, uint64_t image_size)
+void load_random_image(float *image_data_buffer, unsigned long long image_size)
 {
 	int width, height, channels;
 	char image_name[64];
-	snprintf(image_name, 64, "cropped_102flowers/cropped_%04i.png", (rand() + rand() ^ rand()) % IMAGE_COUNT);
+	snprintf(image_name, 64, "cropped_102flowers/cropped_%04u.png", (unsigned)(rand() + rand() ^ rand()) % IMAGE_COUNT);
+
+	//PRIu64
 
 	uint8_t *requested_image = stbi_load(image_name, &width, &height, &channels, 0);
     if (requested_image == NULL) 
@@ -15,18 +17,18 @@ void load_random_image(float *image_data_buffer, uint64_t image_size)
 
 	if (width * height * channels != image_size)
 	{
-		printf("Supplied image size (%llu) does not match loaded image size (%llu).\n", image_size, (uint64_t)width * height * channels);
+		printf("Supplied image size (%llu) does not match loaded image size (%llu).\n", image_size, (unsigned long long)width * height * channels);
         exit(EXIT_FAILURE);
 	}
 
 	uint8_t is_flipped = rand() & 1;
 	if (is_flipped)
 	{
-		for (uint64_t image_row = 0; image_row < height; image_row++)
+		for (unsigned long long image_row = 0; image_row < height; image_row++)
 		{
-			for (uint64_t image_column = 0; image_column < width; image_column++)
+			for (unsigned long long image_column = 0; image_column < width; image_column++)
 			{
-				for (uint64_t image_channel_index = 0; image_channel_index < channels; image_channel_index++)
+				for (unsigned long long image_channel_index = 0; image_channel_index < channels; image_channel_index++)
 				{
 					image_data_buffer[channels * (image_row * width + image_column) + image_channel_index] = 2.0f * (float)requested_image[channels * (image_row * width + width - 1 - image_column) + image_channel_index] / UINT8_MAX - 1.0f;
 				}
@@ -35,7 +37,7 @@ void load_random_image(float *image_data_buffer, uint64_t image_size)
 	}
 	else
 	{
-		for (uint64_t image_data_index = 0; image_data_index < image_size; image_data_index++)
+		for (unsigned long long image_data_index = 0; image_data_index < image_size; image_data_index++)
 		{
 			image_data_buffer[image_data_index] = 2.0f * (float)requested_image[image_data_index] / UINT8_MAX - 1.0f;
 		}
@@ -44,9 +46,9 @@ void load_random_image(float *image_data_buffer, uint64_t image_size)
 	stbi_image_free(requested_image);
 }
 
-void save_image(const char* name, uint64_t width, uint64_t height, int channels, float* data)
+void save_image(const char* name, unsigned long long width, unsigned long long height, int channels, float* data)
 {
-	uint64_t image_size = width * height * channels;
+	unsigned long long image_size = width * height * channels;
 
 	uint8_t *modified_image = (uint8_t *)malloc(image_size);
 	if (modified_image == NULL)
@@ -55,7 +57,7 @@ void save_image(const char* name, uint64_t width, uint64_t height, int channels,
 		exit(EXIT_FAILURE);
 	}
 
-	for (uint64_t image_data_index = 0; image_data_index < image_size; image_data_index++)
+	for (unsigned long long image_data_index = 0; image_data_index < image_size; image_data_index++)
 	{
 		if (data[image_data_index] <= -1.0f)
 		{
@@ -75,7 +77,7 @@ void save_image(const char* name, uint64_t width, uint64_t height, int channels,
 	free(modified_image);
 }
 
-void adjoin_images(float *image_data_buffer, float *image_data_1, float*image_data_2, uint64_t single_image_size)
+void adjoin_images(float *image_data_buffer, float *image_data_1, float*image_data_2, unsigned long long single_image_size)
 {
 	for (int image_data_index_1 = 0; image_data_index_1 < single_image_size; image_data_index_1++)
 	{
@@ -88,14 +90,14 @@ void adjoin_images(float *image_data_buffer, float *image_data_1, float*image_da
 	}
 }
 
-void add_to_adjoined_image(float *image_data_buffer, float *image_data_1, float *image_data_2, uint64_t single_image_width, uint64_t single_image_height, uint64_t single_image_channels, uint64_t start_index)
+void add_to_adjoined_image(float *image_data_buffer, float *image_data_1, float *image_data_2, unsigned long long single_image_width, unsigned long long single_image_height, unsigned long long single_image_channels, unsigned long long start_index)
 {
-	uint64_t single_image_size = single_image_width * single_image_height * single_image_channels;
-	uint64_t stride_bytes = single_image_width * single_image_channels;
+	unsigned long long single_image_size = single_image_width * single_image_height * single_image_channels;
+	unsigned long long stride_bytes = single_image_width * single_image_channels;
 
 	for (int image_data_index = 0; image_data_index < single_image_size * 2; image_data_index++)
 	{
-		uint64_t single_image_data_index = image_data_index % (stride_bytes) + (image_data_index / (stride_bytes * 2)) * stride_bytes;
+		unsigned long long single_image_data_index = image_data_index % (stride_bytes) + (image_data_index / (stride_bytes * 2)) * stride_bytes;
 
 		if (image_data_index % (stride_bytes * 2) < stride_bytes)
 		{
@@ -111,14 +113,14 @@ void add_to_adjoined_image(float *image_data_buffer, float *image_data_1, float 
 /*
 void save_node_network_as_image(struct Node_Network *node_network, char *name)
 {
-	uint64_t image_width = (uint64_t)ceil(sqrt(node_network->node_layers[0].input_count / 3));
-	uint64_t image_height = (uint64_t)ceil(sqrt(node_network->node_layers[0].input_count / 3));
+	unsigned long long image_width = (unsigned long long)ceil(sqrt(node_network->node_layers[0].input_count / 3));
+	unsigned long long image_height = (unsigned long long)ceil(sqrt(node_network->node_layers[0].input_count / 3));
 
-	for (uint64_t node_layer_index = 0; node_layer_index < node_network->node_layer_count; node_layer_index++)
+	for (unsigned long long node_layer_index = 0; node_layer_index < node_network->node_layer_count; node_layer_index++)
 	{
-		uint64_t individual_image_side;
+		unsigned long long individual_image_side;
 
-		individual_image_side = (uint64_t)ceil(sqrt(node_network->node_layers[node_layer_index].output_count / 3));
+		individual_image_side = (unsigned long long)ceil(sqrt(node_network->node_layers[node_layer_index].output_count / 3));
 		image_height += individual_image_side;
 
 		if (image_width < individual_image_side)
@@ -134,26 +136,26 @@ void save_node_network_as_image(struct Node_Network *node_network, char *name)
 		exit(EXIT_FAILURE);
 	}
 
-	for (uint64_t image_data_index = 0; image_data_index < image_width * image_height * 3; image_data_index++)
+	for (unsigned long long image_data_index = 0; image_data_index < image_width * image_height * 3; image_data_index++)
 	{
 		image_data[image_data_index] = -2.0f;
 	}
 
-	uint64_t next_layer_image_start_index = 0;
+	unsigned long long next_layer_image_start_index = 0;
 
 	{
-		uint64_t input_layer_image_side = (uint64_t)ceil(sqrt(node_network->node_layers[0].input_count / 3));
-		for (uint64_t image_data_index = 0; image_data_index < node_network->node_layers[0].input_count; image_data_index++)
+		unsigned long long input_layer_image_side = (unsigned long long)ceil(sqrt(node_network->node_layers[0].input_count / 3));
+		for (unsigned long long image_data_index = 0; image_data_index < node_network->node_layers[0].input_count; image_data_index++)
 		{
 			image_data[(image_data_index / input_layer_image_side) * image_width + (image_data_index % input_layer_image_side)] = node_network->node_layers[0].inputs[image_data_index];
 		}
 		next_layer_image_start_index = image_width * 3 * input_layer_image_side;
 	}
 
-	for (uint64_t node_layer_index = 0; node_layer_index < node_network->node_layer_count; node_layer_index++)
+	for (unsigned long long node_layer_index = 0; node_layer_index < node_network->node_layer_count; node_layer_index++)
 	{
-		uint64_t individual_image_side = (uint64_t)sqrt(node_network->node_layers[node_layer_index].output_count / 3);
-		for (uint64_t image_data_index = 0; image_data_index < node_network->node_layers[node_layer_index].output_count; image_data_index++)
+		unsigned long long individual_image_side = (unsigned long long)sqrt(node_network->node_layers[node_layer_index].output_count / 3);
+		for (unsigned long long image_data_index = 0; image_data_index < node_network->node_layers[node_layer_index].output_count; image_data_index++)
 		{
 			//image_data[next_layer_image_start_index + image_data_index] = node_network->node_layers[0].inputs[(image_data_index / image_width) * individual_image_side + (image_data_index % (image_width))];
 			image_data[next_layer_image_start_index + (image_data_index / individual_image_side) * image_width + (image_data_index % individual_image_side)] = node_network->node_layers[node_layer_index].raw_linear_outputs[image_data_index];
